+++ packages/agtree/test/converter/network/network-rule-converter.test.ts
@@ -342,4 +342,101 @@ describe('NetworkRuleConverter', () => {
             }).toThrow(new RuleConversionError(expected));
         });
     });
+
+    describe('convertToUbo', () => {
+        // Based on https://github.com/AdguardTeam/Scriptlets/blob/673d48442bb43c7e32caeaeee13a33f94985a9a2/tests/api/index.spec.js#L685-L781
+        test.each([
+            {
+                actual: '||example.com^$script,redirect=noopjs:99',
+                expected: ['||example.com^$script,redirect=noop.js'],
+            },
+            {
+                actual: '||example.com^$xmlhttprequest,redirect=nooptext',
+                expected: ['||example.com^$xhr,redirect=noop.txt'],
+            },
+            {
+                actual: '||example.com/*.css$important,redirect=noopcss',
+                expected: ['||example.com/*.css$important,redirect=noop.css,stylesheet'],
+            },
+            {
+                // image type is supported by nooptext too
+                actual: '||example.com^$image,redirect=nooptext',
+                expected: ['||example.com^$image,redirect=noop.txt'],
+            },
+            {
+                // eslint-disable-next-line max-len
+                actual: '||example.com/images/*.png$image,important,redirect=1x1-transparent.gif,domain=example.com|example.org',
+                // eslint-disable-next-line max-len
+                expected: ['||example.com/images/*.png$image,important,redirect=1x1.gif,domain=example.com|example.org'],
+            },
+            {
+                actual: '||example.com/vast/$important,redirect=empty,~third-party',
+                expected: ['||example.com/vast/$important,redirect=empty,~3p'],
+            },
+            {
+                // add source type modifiers while conversion if there is no one of them
+                // eslint-disable-next-line max-len
+                actual: '||example.com/images/*.png$redirect=1x1-transparent.gif,domain=example.com|example.org,important',
+                // eslint-disable-next-line max-len
+                expected: ['||example.com/images/*.png$redirect=1x1.gif,domain=example.com|example.org,important,image'],
+            },
+            {
+                actual: '||example.com/*.mp4$important,redirect=noopmp4-1s,~third-party',
+                expected: ['||example.com/*.mp4$important,redirect=noop-1s.mp4,~3p,media'],
+            },
+            {
+                actual: '||example.com/*.css$important,redirect=noopcss',
+                expected: ['||example.com/*.css$important,redirect=noop.css,stylesheet'],
+            },
+            {
+                actual: '||ad.example.com^$redirect=nooptext,important',
+                // eslint-disable-next-line max-len
+                expected: ['||ad.example.com^$redirect=noop.txt,important,image,media,subdocument,stylesheet,script,xhr,other'],
+            },
+            {
+                // eslint-disable-next-line max-len
+                actual: '||imasdk.googleapis.com/js/sdkloader/ima3.js$script,important,redirect=google-ima3,domain=example.org',
+                // eslint-disable-next-line max-len
+                expected: ['||imasdk.googleapis.com/js/sdkloader/ima3.js$script,important,redirect=google-ima.js,domain=example.org'],
+            },
+            // $redirect-rule
+            {
+                actual: '||example.com^$xmlhttprequest,redirect-rule=nooptext',
+                expected: ['||example.com^$xhr,redirect-rule=noop.txt'],
+            },
+            {
+                // image type is supported by nooptext too
+                actual: '||example.com^$image,redirect-rule=nooptext',
+                expected: ['||example.com^$image,redirect-rule=noop.txt'],
+            },
+            {
+                // eslint-disable-next-line max-len
+                actual: '||example.com/images/*.png$image,important,redirect-rule=1x1-transparent.gif,domain=example.com|example.org',
+                // eslint-disable-next-line max-len
+                expected: ['||example.com/images/*.png$image,important,redirect-rule=1x1.gif,domain=example.com|example.org'],
+            },
+            {
+                actual: '||example.com/vast/$important,redirect-rule=empty,~third-party',
+                expected: ['||example.com/vast/$important,redirect-rule=empty,~3p'],
+            },
+            {
+                // add source type modifiers while conversion if there is no one of them
+                // eslint-disable-next-line max-len
+                actual: '||example.com/images/*.png$redirect-rule=1x1-transparent.gif,domain=example.com|example.org,important',
+                // eslint-disable-next-line max-len
+                expected: ['||example.com/images/*.png$redirect-rule=1x1.gif,domain=example.com|example.org,important,image'],
+            },
+            {
+                actual: '||example.com/*.mp4$important,redirect-rule=noopmp4-1s,~third-party',
+                expected: ['||example.com/*.mp4$important,redirect-rule=noop-1s.mp4,~3p,media'],
+            },
+            {
+                actual: '||ad.example.com^$redirect-rule=nooptext,important',
+                // eslint-disable-next-line max-len
+                expected: ['||ad.example.com^$redirect-rule=noop.txt,important,image,media,subdocument,stylesheet,script,xhr,other'],
+            },
+        ])('should convert \'$actual\' to \'$expected\'', (testData) => {
+            expect(testData).toBeConvertedProperly(NetworkRuleConverter, 'convertToUbo');
+        });
+    });
 });