+++ packages/agtree/test/smoke/cjs/index.js
@@ -1,0 +1,8 @@
+const { RuleParser } = require('@adguard/agtree');
+const { ok } = require('assert');
+
+const ruleNode = RuleParser.parse('||example.com^');
+
+ok(ruleNode);
+
+console.log('Smoke test passed');