+++ packages/css-tokenizer/rollup.config.ts
@@ -10,9 +10,6 @@ import resolve from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
 import dtsPlugin from 'rollup-plugin-dts';
 import json from '@rollup/plugin-json';
-import terser from '@rollup/plugin-terser';
-import nodePolyfills from 'rollup-plugin-polyfill-node';
-import { getBabelOutputPlugin } from '@rollup/plugin-babel';
 import path from 'path';
 import { readFileSync } from 'fs';
 
@@ -22,7 +19,7 @@ const BASE_NAME = 'CSSTokenizer';
 const BASE_FILE_NAME = 'csstokenizer';
 const PKG_FILE_NAME = 'package.json';
 
-const distDirLocation = path.join(ROOT_DIR, 'dist');
+const distDir = path.join(ROOT_DIR, 'dist');
 const pkgFileLocation = path.join(ROOT_DIR, PKG_FILE_NAME);
 
 // Read package.json
@@ -43,150 +40,61 @@ for (const field of REQUIRED_PKG_FIELDS) {
 }
 
 // Generate a banner with the current package & build info
-const BANNER = `/*
+const banner = `/*
  * ${BASE_NAME} v${pkg.version} (build date: ${new Date().toUTCString()})
  * (c) ${new Date().getFullYear()} ${pkg.author}
  * Released under the ${pkg.license} license
  * ${pkg.homepage}
  */`;
 
-// Pre-configured TypeScript plugin
-const typeScriptPlugin = typescript({
-    compilerOptions: {
-        // Don't emit declarations, we will do it in a separate command "pnpm build-types"
-        declaration: false,
-    },
-});
-
-const terserPlugin = terser({
-    sourceMap: true,
-    output: {
-        // Keep the banner in the minified output
-        preamble: BANNER,
-    },
-    compress: {
-        // It has negative impact on performance, so we disable it
-        reduce_funcs: false,
-    },
-});
-
 // Common plugins for all types of builds
 const commonPlugins = [
     json({ preferConst: true }),
     commonjs({ sourceMap: false }),
     resolve({ preferBuiltins: false }),
-    typeScriptPlugin,
-];
-
-// Plugins for Node.js builds
-const nodePlugins = [
-    ...commonPlugins,
-];
-
-// Plugins for browser builds
-const browserPlugins = [
-    ...commonPlugins,
-    nodePolyfills(),
-    // Provide better browser compatibility with Babel
-    getBabelOutputPlugin({
-        presets: [
-            [
-                '@babel/preset-env',
-                {
-                    targets: {
-                        // https://github.com/browserslist/browserslist#best-practices
-                        browsers: [
-                            'last 1 version',
-                            '> 1%',
-                            'not dead',
-
-                            // Specific versions
-                            'chrome >= 88',
-                            'firefox >= 84',
-                            'edge >= 88',
-                            'opera >= 80',
-                            'safari >= 14',
-                        ],
-                    },
-                },
-            ],
-        ],
-        allowAllFormats: true,
-        compact: false,
+    typescript({
+        tsconfig: path.join(ROOT_DIR, 'tsconfig.json'),
+        compilerOptions: {
+            incremental: true,
+            declaration: true,
+            declarationDir: path.join(ROOT_DIR, 'dist/types'),
+        },
+        include: ['./src/**/*.ts'],
+        exclude: ['./node_modules', './test'],
+        outputToFilesystem: true,
     }),
-    // Minify the output with Terser
-    terserPlugin,
 ];
 
 // CommonJS build configuration
-const cjs = {
-    input: path.join(ROOT_DIR, 'src', 'index.ts'),
+const node = {
+    input: path.join(ROOT_DIR, 'src/index.ts'),
     output: [
         {
-            file: path.join(distDirLocation, `${BASE_FILE_NAME}.js`),
+            file: path.join(distDir, `${BASE_FILE_NAME}.js`),
             format: 'cjs',
             exports: 'auto',
             sourcemap: false,
-            banner: BANNER,
+            banner,
         },
-    ],
-    plugins: nodePlugins,
-};
-
-// ES module build configuration
-const esm = {
-    input: path.join(ROOT_DIR, 'src', 'index.ts'),
-    output: [
         {
-            file: path.join(distDirLocation, `${BASE_FILE_NAME}.esm.mjs`),
+            file: path.join(distDir, `${BASE_FILE_NAME}.mjs`),
             format: 'esm',
-            exports: 'auto',
             sourcemap: false,
-            banner: BANNER,
+            banner,
         },
     ],
-    plugins: nodePlugins,
+    plugins: commonPlugins,
 };
 
-// Browser-friendly UMD build configuration
-const umd = {
-    input: path.join(ROOT_DIR, 'src', 'index.ts'),
-    output: [
-        {
-            file: path.join(distDirLocation, `${BASE_FILE_NAME}.umd.min.js`),
-            format: 'umd',
-            sourcemap: false,
-            banner: BANNER,
-            name: BASE_NAME,
-        },
-    ],
-    plugins: browserPlugins,
-};
-
-// Browser-friendly IIFE build configuration
-const iife = {
-    input: path.join(ROOT_DIR, 'src', 'index.ts'),
-    output: [
-        {
-            file: path.join(distDirLocation, `${BASE_FILE_NAME}.iife.min.js`),
-            format: 'iife',
-            exports: 'auto',
-            sourcemap: false,
-            banner: BANNER,
-            name: BASE_NAME,
-        },
-    ],
-    plugins: browserPlugins,
-};
-
-// Merge .d.ts files (requires `tsc` to be run first, because it merges .d.ts files from `dist/types` directory)
+// Merge .d.ts files (requires `tsc` to be run first,
+// because it merges .d.ts files from `dist/types` directory)
 const dts = {
-    input: path.join(ROOT_DIR, 'dist', 'types', 'src', 'index.d.ts'),
+    input: path.join(ROOT_DIR, 'dist/types/src/index.d.ts'),
     output: [
         {
-            file: path.join(distDirLocation, `${BASE_FILE_NAME}.d.ts`),
+            file: path.join(distDir, `${BASE_FILE_NAME}.d.ts`),
             format: 'es',
-            banner: BANNER,
+            banner,
         },
     ],
     plugins: [
@@ -195,4 +103,4 @@ const dts = {
 };
 
 // Export build configs for Rollup
-export default [cjs, esm, umd, iife, dts];
+export default [node, dts];