+++ packages/agtree/CHANGELOG.md
@@ -8,18 +8,48 @@ The format is based on [Keep a Changelog][keepachangelog], and this project adhe
 [keepachangelog]: https://keepachangelog.com/en/1.0.0/
 [semver]: https://semver.org/spec/v2.0.0.html
 
-## Unreleased
+## [2.1.0] - 2024-09-18
+
+### Added
+
+- `UboParameterListParser` to parse uBO parameter lists, like scriptlet parameters.
+
+### Changed
+
+- Scriptlet injection parser now requires quotes for AdGuard scriptlets and uses `UboParameterListParser` for uBO
+  scriptlets to provide more consistent parsing.
+
+[2.1.0]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/agtree-v2.1.0
+
+## [2.0.2] - 2024-09-10
+
+### Fixed
+
+- Handling last position in `PositionProvider`.
+
+### Removed
+
+- Browser-specific builds. If you need to use AGTree in browser environment, you can use jsDelivr's automatic builds
+  or bundlers like Webpack / Rollup / ESBuild without any issues, AGTree does not have any non-browser-compatible code.
+
+[2.0.2]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/agtree-v2.0.2
+
+## [2.0.1] - 2024-09-06
 
 ### Added
 
 - Support for uBO's legacy `script:inject` mask.
 - Support for uBO's redirect priority in the compatibility table name normalization.
+- `resource_types` field to the redirects compatibility table.
+- Support for converting network rules to uBO syntax.
 
 ### Fixed
 
 - Conversion for uBO's `rc` and `ra` scriptlets.
 - Converter now unescape scriptlet separators from the previous rule, if needed (e.g. it unescape commas when converting
   from uBO to AG syntax).
+
+[2.0.1]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/agtree-v2.0.1
 
 ## [2.0.0] - 2024-08-15